Use Rlock when Get