@kevburnsjr 2021-09-18
I just noticed this comment in math/rand. RLock might not be appropriate.
```
// The default Source is safe for concurrent use by multiple goroutines, but
// Sources created by NewSource are not.
```
https://cs.opensource.google/go/go/+/refs/tags/go1.17.1:src/math/rand/rand.go;l=12