@sean-public 2021-09-18
> Parallel bench test shows latency and throughput improvements of RLock are very significant for concurrent access.

If the "new times" listed are using RLock, there's notable performance slowdowns in `IncGet-12` and `DecGet-12`. I know there's no synthetic benchmark for this already extant and, further, all synthetic benchmarks have serious shortcomings, but I believe that in a very common workload with a blend of, say, 20% writes and 80% reads all in parallel, there would be worse performance for writes because they are waiting for all of the pending read locks to free.

Just thinking out loud here 🤔