@kevburnsjr 2021-09-19
@sean-public You might actually be right about that...
https://gist.github.com/kevburnsjr/a66c19a16d2ae9d74cd501a8b9b3c6b4
```
cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
BenchmarkIncSet-12               4232509               277.1 ns/op      15272328.16 MB/s              61 B/op          3 allocs/op
BenchmarkIncGet-12               8680887               163.1 ns/op      53221254.06 MB/s               0 B/op          0 allocs/op
BenchmarkDecSet-12               7107218               179.5 ns/op      39584672.41 MB/s              61 B/op          3 allocs/op
BenchmarkDecGet-12               8546569               166.6 ns/op      51298366.60 MB/s               0 B/op          0 allocs/op
BenchmarkParallel/readpct-0-12           3960511               307.9 ns/op             1 B/op          1 allocs/op
BenchmarkParallel/readpct-20-12          4383126               291.3 ns/op             0 B/op          0 allocs/op
BenchmarkParallel/readpct-50-12          4119375               313.9 ns/op             0 B/op          0 allocs/op
BenchmarkParallel/readpct-80-12          3813206               340.5 ns/op             0 B/op          0 allocs/op
BenchmarkParallel/readpct-100-12        33822919                37.14 ns/op            0 B/op          0 allocs/op
PASS
ok      fast-skiplist   27.501s
```